+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 There are no releases, just rolling latest master. The list is reverse chronologic order
 (newest entries first).
 
+- `submitobs.py` now uses ArgParse (#9)
 - Each observation has its own log file now (using logging module)
 - `station plan --show` command now shows frequency as extra column
 - More robust local observations management. It's now possible to print observations